Use bamboo_phoenix to leverage templating
