+++ apps/shoutouts_web/lib/shoutouts_web/templates/layout/email.html.eex
@@ -1,0 +1,8 @@
+<html>
+  <head>
+    <link rel="stylesheet" href="<%= Routes.static_url(ShoutoutsWeb.Endpoint, "/css/app.css") %>">
+  </head>
+  <body>
+    <%= render @view_module, @view_template, assigns %>
+  </body>
+</html>
