+++ apps/shoutouts_web/lib/shoutouts_web/views/email_view.ex
@@ -1,0 +1,3 @@
+defmodule ShoutoutsWeb.EmailView do
+  use ShoutoutsWeb, :view
+end
